Xrootd: Update logging to log4j.
This is the first in a series of patches to clean up the xrootd door.
This patch moves logging from classic cells say/esay and in places
printing to System.err or System.out to log4j. Log messages are not
changed. I have not spent much time figuring out which log level to use.
In general I have mapped say->info and esay->warn, and System.out to
info or debug and System.err to error or warn.

Target: trunk
Target: 1.9.2
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/263/