migration module: Fix pool selection in case pool is down
The cost computation in the migration module is wrong in case a pool in
the target pool group or link is down. In that case the cost is
infinite, but with a cost factor of zero, the resulting cost becomes NaN
(0 times infinity is NaN).

This patch fixes this by adding an explicit check for a cost factor of zero.

I have added two toString() methods that I used during debugging.

A unit test has been added to verify that POSITIVE_INFINITY is now
handled correctly. This required making some of the classes public, as
we put the unit tests in a different directory.

Target: trunk
Target: 1.9.2
Target: 1.9.1
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/268/