PnfsManager: Move checksum conflict check into provider
To avoid that checksums are accidentally changed (something that should
never happen during normal operation since dCache is essentially a
WORM), PnfsManager would reject changing an existing checksum. To
enforce this it had to query the existing checksum from the name space
provider. This incurred an extra PNFS read which could be avoided as the
name space provider already had to read the existing checksums when
writing the new value.

This patch moves the check into the name space provider. Thus for PNFS
we can perform the check in the read-modify-write cycle and thus avoid
an extra read. For Chimera there is no difference.

This patch is submitted in response to Timur's analysis of changes in
PNFS load in 1.9.2.

Target: 1.9.2
Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/280/