Restart upon fatal VM error.
This patch modifies the uncaught exception handler such that the domain
is shut down when a VirtualMachineError is caught. This may for instance
happen when an OutOfMemoryError occurs. The dCache wrapper script will
automatically restart the domain.

Notice that many components catch Throwable and hence the exception
never reaches the UncaughtExceptionHandler. For instance
ThreadPoolExecutor and FutureTask do this. We should be aware of this
when writing code and modify our exception handling such that these
fatal errors are propagated.

The patch removed CellThreadGroup since its only value was debug logging
when finalize() was called. The standard thread group implementation
will forward uncaught exceptions to the default uncaught exception
handler.

I added a new Logger to SystemCell such that the error messages from the
uncaught exception handler could be logged at an approriate log level.
In a future patch other log statements in SystemCell should be updated
to use the new logger.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/285/