Add support for an option to tell comparator to continue when faced with an inconsistency.
There has been a request for a new feature to the comparator to allow it to continue when it encounters an 
inconsistency between PNFS and Chimera, rather than failing.  This patch supports this by adding a command-line 
option (-k).

The default behaviour of Comparator is still to fail at the first error.  Specifying -k tells the comparator to 
attempt all PNFS-IDs irrespective of how many fail.

http://rb.dcache.org/r/193/

Acked-by: Gerd