Stream LIST output while it is produced
The FTP list command used to generate the output in a StringBuffer. This
had two negative effects: For large directories is used a lot of memory,
and it took a long time before the first files were listed. This patch
refactors the list command such that the output is sent as it is
generated.

The patch also refactors the overall structure of list operations:
Socket creation and shutdown is now shared with NLST, and the list
method is embedded in ac_list.

Target: trunk
Ticket: #4493
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/241/