PoolManager: Discard requests when TTL has been exceeded
Doors and other 'clients' of PoolManager often retry requests when
PoolManager doesn't answer within a certain amount of time. Often the
request is resubmitted. The PoolManager on the other doesn't reply until
the file is ready for transfer, which may take an arbitrary amount of
time if the file is on a pool that is not online or when it is on tape.

The resubmitted request is enqueued alongside the original request. This
causes the amount of requests to queue up, and on a busy system this may
cause a significant amount of memory to be consumed in the PoolManager.

To counter this problem the patch uses the TTL support in cell messages
to discard requests once the TTL has been exceeded. Discarding the
message does not alter the behavior, as the door no longer waits for
the reply at this time.

Notice that we do not dequeue the activity (staging, p2p, etc). We only
drop the request message to conserve memory. Often the door will
resubmit the request and it will be attached to the same handler. We may
later consider changing the retry logic of the PoolManager such that we
do not retry if there are no requests attached, however that change is
too complicated to do on 1.9.2 and should be done in a different patch.

I submit this for inclusion in both trunk and 1.9.2 and 1.9.1. Since FZK
suffers from out of memory problems, Patrick requested we make this change
in stable branches too.

Notice the change in addRequest(). Previously it was possible to detect
if the handler was uninitialized by looking at the number of attached
messages. This is no longer possible as we discard messages when the TTL
is exceeded. Instead we now look at _pnfsFileLocation to determine if
the handler is uninitialized.

Target: trunk
Target: 1.9.2
Target: 1.9.1
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/257/