Pool: Fix sticky flag expiration during pool startup
MetaDataRecord provides a isSticky() method. The old implementation
would change the value of isSticky() as sticky flags would expire. This
caused problems during pool startup, as isSticky() is called several
times and if the value changes in the middle, accounting of removable
files would be wrong.

I could have worked around the issue by changing the startup procedure.
The MetaDataRecord becomes however simpler if we change the definition
of isSticky() such that it is not timing dependant. That makes it more
predictable and easier to use.

I am not sure to what extend the bug is present in pre-trunk versions.
Since we have not received reports about this problem (and a simple
restart is enough to fix it should it happen), I suggest we limit this
patch to trunk (trunk is rather different from previous releases).

Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/275/