PnfsManager: Don't sleep on setFileMetaData in case of error.
In 1.9.2 we added a workaround for a race condition in PNFS. Earlier 
releases simply ignored the error that was generated due to the race 
condition, but 1.9.2 tried to be smarter and distinguish between the 
harmless race condition and the real error. However the workaround 
caused request queuing in PnfsManager as the race condition seemed to be 
much more common than believed.

This patch changes the logic to once again ignore the error. An entry is 
added to the log, which will alove us to asses how common the problem 
is.

Target: trunk
Patch: http://rb.dcache.org/r/290
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>