Refactors the dcache script
This patch refactors the dcache script. It

- moves the bulk of the script to shell libraries in share/lib/
- renames several functions to better reflect what they do
- moves more functions to returning values in RET (but not all
  since for some it is simply more convenient to use stdout).
- adds better error handling to several functions.
- refactors the name space mount logic and now also supports chimera.
- adds a SERVICES parameter to node_config.
- moves pid files to /var/run
- treats config/*Setup files as service specific configuration
- defines default paths in share/lib/paths.sh

The old NODE_TYPE types and service flags are still supported, but the
recommended approach is now to define SERVICES as a list of services to
run on the host. There is thus no need to define new node_config flags
whenever we add a new domain type.

NODE_TYPE=admin prepulates the SERVICES field with the services we
typically run on the head node.

Both services and domain names can be listed in SERVICES, making it
possible to only start some instances of a particular service.

The name space mount logic exclusively looks at the list of services
(after merging SERVICES with predefined services according to NODE_TYPE)
and the NAMESPACE type. It no longer tests explicitly on NODE_TYPE.

The name space mount logic now also respects the pnfs setting in
dCacheSetup (we should maybe hard code the default value used in
pnfs.batch in case pnfs is not defined in dCacheSetup).

The name space logic now looks for .(tags)() in the mount point to
determine if PNFS is mounted. Previously a combination of df and grep
was used.

Config file parsing was changed such that node_config parameters are
read into environment variables. These variables are prefixed with
NODE_CONFIG_. This is considerably faster than repeatedly reading the
same configuration file whenever a value is needed. The config.sed sed
script is used to normalise configuration files. Notice that config.sed
should not be considered bullet proof against shell command injection.
There is no point in trying to make it bullet proof, since the values
are used throughout the shell script afterwards and may contain quotes.
We would have to completely disallow the use of quotes in values.

/var/run on both Linux and Solaris is automatically cleaned during boot.
For this reason the LSB dictates that PID files are supposed to be
stored in /var/run. Storing the PID files in /var/run solves the common
problem that dCache doesn't start after an OS crash or power failure.

This patch introduces the configuration parameter pidDir in *Setup.
pidDir defaults to /var/run. Pid files are stored in $pidDir using the
name dcache.<domain>.pid, where <domain> is the cell domain name.

The output of 'dcache status' was changed to list service name for each
domain. It no longer lists the log file location. 'dcache services' now
lists log file location.

TODO: The patch does not update the install.sh script. That script
should be updated to respect SERVICES, and preferably use the same
library scripts.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Owen Synge <owen.synge@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/131/