Support RFC 3659 MLST and MLSD commands
RFC 3659 defines the MLST and MLSD commands. MLST provides information
about a file or a directory. MLSD provides information about the files
in a directory. In contrast to the LIST command, the output of MLST and
MLSD is well-defined and machine readable. The list of facts provided
for a file is negotiable throught the OPTS MLST command.

By default the PERM fact is not included in the output, as it is
expensive to compute the value through the current PermissionHandler
interface.

FYI: The ARC clients utilise the MLSD command to obtain a file listing.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>