jglobus: normalize DN in signing policy matching
Paul discovered a regression when going from JGlobus 1.4 to 1.6. The
issue was that JGlobus 1.6 enforced the CA signing policy. While doing
so it omitted to normalize email addresses that appear in the DN. As
there are several ways of generating a DN containing an email address,
not normalizing the DN may lead to certificates being rejected when they
should have been accepted.

This patch copies a DN normalization method from the JGlobus GridMap
class to SigningPolicyParser. I decided to make a copy, as the version
in GridMap ignores capitalization. While copying the code I changed it
such that the capitalization is respected.

I added two calls to the normalization method. One is when matching the
DN of the CA itself, and the other is when determining whether a DN
matches the pattern given in the signing policy file. In both cases we
transform Email= and E= to emailAddress=.

Notice that the bulk of the normalization method is copied from GridMap.
I am sure there are many other ways of implementing it, but I preferred
to keep the structure of the code already present.

I also chose to ignore the obvious violation of the DRY principle. I
believe we should submit the patch to the JGlobus team, but I don't
expect it to be included literally. The goal of my patch is to
temporarily fix the JGlobus library for our needs. I expect the JGlobus
team to investigate the problem themselves and make a suitable long term
fix.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/320/