xrootd: Collapse double slashes (again)
Patch 304 added code to collapse repeated slashes in xrootd paths.
Unfortunately this could lead to a potential regression in the open
call: The path may be a LFN which is mapped to a PFN while parsing the
token included in the open call. Collapsing double slashes in the LFN
could break this translation.

This patch reverts patch 304 and moves the collapse code to
XrootdDoorListener.

Target: 1.9.1
Target: 1.9.2
Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/321/