With PNFS, avoid unnecessary level-2 reads when looking up a file's metadata.
The getFileMetaData() method returns information about a file's (or
directory's) filesystem metadata.  This includes the file's size.
With PNFS-based filesystems, there is a 2GiB limit on file sizes due
to a corresponding NFSv2 limitation.  To work around this
limitation,the file size is also stored in the level-2 StorageInfo
metadata as the 'l' flag.  If a file size cannot be set correctly, it
is assigned a place-holder value of "1".

Current semantics are that if the file size (as recorded in the
filesystem) is greater than 1 then it is used.  If the size is 1 or
less then PnfsManager will attempt to parse the level-2 StorageInfo
for the 'l' flag.  Should this be unsuccessful (for whatever reason)
then the file size will be used.

The issue is that the level-2 metadata is always fetched whether it is
needed or not.  If the file size in the filesystem is greater than 1
then the resulting value will be thrown away.

This patch addresses this issue by only reading the level-2
StorageInfo if the file length is 1 or less.