scripts: Check pnfsManager variable for custom node type
On nodes with node type 'custom', the pnfsManager variable as defined
in node_config determines if the domain containing PnfsManager should
be started. This patch fixes a bug that caused PnfsManager to always be
started on a node with node type 'custom'.

Target: trunk
Target: 1.9.3
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/332/