info-provider: fix the Version property value for SRM GlueControlProtocol objects
According to the document "Usage of Glue Schema v1.3 for WLCG Installed Capacity information" document, the GLUE
ControlProtocol objects that represent SRM endpoints must have a GlueControlProtocolVersion attribute that is either
'1.1.0' or '2.2.0'.  The current info provider publishes '2.2', so does not conform to this requirement.

RB: http://rb.dcache.org/r/331/
Acked-by: Tigran