Allow "pool create" to work on non-existing directories.
Previously "pool create" would fail with an ugly error message if the
parent directory of the new pool did not exist. This patch changes the
logic such that the parent directory is created before we check the
amount of free space.

This change was requested by Patrick.

Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/168/