pool: Refactors the migration server and first part of checksum handling
The migration module achieves idempotens by locating existing copies of
a file and skipping the copy if found. Alex suggested that we add an
option to force a checksum computation on the existing copy to verify
its integrity. This is in particular relevant when the source copy is to
be destroyed.

This change means that the "update" of the existing copy is no longer a
near instantaneous operation, as computing the checksum may take quite a
while. Hence we have to execute it in a different thread, we have to
support pinging as we cannot put a fixed upper bound on the processing
time, and we have to be able to cancel it. For a regular copy we already
support these operations. This patch refactors the migration server such
that the PoolMigrationCopyReplica message is used for both copying the
file and for updating an existing copy (depending on whether the file is
new or already exists).

At the same time I replace the taskId field with a proper UUID. As UUIDs
are supposed to be unique, there is no reason to qualify the task ID
with the pool name. We also solve potential issue with pool restarts,
which would reset the task ID counter.

As these changes affect the pool communication, we cannot implement
everything in one release. This patch adds the migration server side
implementation and extends the messages with the new fields. The
migration server is able to identify tasks both by task ID and UUID,
thus making it compatible with both old and new pools. In a future
release we will remove the task ID part and let the migration module use
the PoolManangerCopyReplica message for both copies and updates. In a
release after that release we can remove the deprecated
PoolMigrationUpdateReplica message.

Target: trunk
Jira: https://project.nordu.net/browse/NDGF-191
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/396/