Add saxon jar file
Prior to 1.9.3 being cut, saxon file was added as an alternative XSLT processor for the info system.  This requires the saxon.jar file to be present.

Unfortunately, the jar file was added only to the 1.9.2 branch and not trunk.  This patch is to add the saxon jar file to /opt/d-cache/modules/external/saxon/saxon.jar (not visible within the patch) and adjust the ant build system to include this jar file.

Previously, the jar file was added as /opt/d-cache/classes/saxon.jar.  This is different from how other 3rd party jar files are stored.  The others are all installed in a separate directory.  To fix this inconsistency, this patch now installs the saxon jar file in a separate directory (/opt/d-cache/classes/saxon) and updates the wrapper script to allow the directory of the saxon.jar file to be configured with the default location being /opt/d-cache/classes/saxon/

This patch is for 1.9.3, 1.9.4 and trunk.

RB: http://rb.dcache.org/r/402/
Acked-by: Tigran