gPlazma: Change the AuthzQueryHelper class to receive CellEndpoint's instead of CellAdapter's.
AuthzQueryHelper receives in its helper methods the CellAdapter which is resposible to sending and receiving the reply to a mapping of X509Certificate to a uid and gid. This makes it impossible to use it with cell components created using the Spring framework as they are not exposing the CellAdapter but only the CellEndpoint interface. This patch changes the type signature of the methods to receive a CellEndpoint.

Target: trunk
Patch: http://rb.dcache.org/r/398/
Acked-by: Tigran