pnfsmanager: Merge Provider interfaces
In the latest revision, the distinction between the three provider
intefaces becomes less clear. This patch merges all three interfaces
into a single NameSpaceProvider interface. This is also a prerequisite
to

a) updating this interface to use a Subject field on all methods
b) reinventing PnfsHandler to use the same or a similar interface

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/435/