script: Add --force option to dump heap command
When the JVM hangs or is slow to respond, dumping the heap with jmap
often fails. This patch adds a --force option to the init script which
in turn will cause jmap to be called with the -F option. Then the dump
should succeed even with the JVM is unresponsive.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/422/