cells: Add workaround for Java bug 4509080
There is a bug in NIO: When using the streams API on a socket channel,
those streams are not fully bidirectional: A read can block a write.
Hence a read thread blocked on a read call can block all attempts to
write to that channel.

Bug 4774871 suggest a workaround for this bug. This patch adds the
workaround to DummyStreamEngine.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/401/