pnfsmanager: Add decorator of NameSpaceProvider for permission handling
This patch add PermissionNameSpaceProvider, a decorator for
NameSpaceProvider. The current implementation checks Unix permissions
for a few operations corresponding to create, delete, rename, read, and
lookup.

The current implementation is incomplete, as it lacks ACL support. It
can wrap both BasicNameSpaceProvider and ChimeraNameSpaceProvider,
although the intention for Chimera is to move the checks into the name
space provider, as Chimera can optimize data retrival from the database.

We must also consider to what extend to add permission checking for all
other operations. We certainly cannot check permissions for every small
operation as it would be too expensive. This should however be a
consideration of the client making the request (e.g. the door). It
should decide to what extend being granted one operation implies access
to other operations. E.g. it is reasonable to assume that if a user was
allowed to create an entry, then that user is also allowed to set
checksum and file size of that file (assuming the provider implements
sane permission handling).

The current implementation serves as a starting point. It allows us to
get the important operations checked, meaning we can start working on
the doors.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/442/