LoginBroker: 'ls' by multiple protocols
The 'ls' command of LoginBroker accepts as argument a protocol which makes 'ls' return only doors of that protocol. This patch extends this behaviour by accepting not one protocol but a list of comma separated protocols. Particularly useful when one wants to get the secure and non-secure versions of a protocol in one go.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/460/