core: Use EnumSet (bit mask) to represent a set of file attributes
Using an EnumSet (essentially a bit mask) is faster and more compact. As
an added benefit we can add checks to FileAttributes to ensure that
fields are indeed defined when once tries to extract them.

This is a replacement for Tigran's patch 478.

Target: trunk
Patch: http://rb.dcache.org/r/482/
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>