pool: Fixed a race condition in migration module
When multiple migration jobs run at the same time, either on the same
pool, or on two different pools, where one is among the target set of
the other, then there are obvious race conditions in the system. These
race conditions exist at the "migration job level". They are "created"
by the user, in that the user defined migration jobs that act upon the
same files. There is nothing we can do prevent such race conditions
(similar to how your favorit programming language cannot prevent you
from creating race conditions in your programs).


Each task (transfer) of a migration job should ideally be atomic. We
don't have a distributed transaction manager in dCache, so we cannot
make it truly atomic in case of communication failures, pool crashes,
etc. However if we got multible migration jobs, A and B, acting upon the
same file F, then either A should do its task on F first or B should do
it first. The outcome is still non-deterministic depending on which job
comes first, however the set of possible outcomes should be limitted to
these two cases. Unfortunately this is not the case in the current
system.


We don't have the infra structure in the pool yet to address this
problem in general. This patch addresses one particularly nasty
incarnation of this problem. The problem is easy to trigger. On pool0
you do 'migration move -permanent pool1' and on pool1 you do 'migration
move -permanent pool0'. In the current versions, this will result in
data loss. The problem is that if pool0 moves a file A to pool1, then as
soon as A is complete on pool1, the migration job on pool1 will pick it
up and attempt to move it to pool0. If the task on pool0 that moved A to
pool1 is not done yet at this time, the pool1 will observe that pool0
already got a copy and complete the move by deleting its local copy.
That is then followed by pool0 completing the migration task and also
deleting its copy. Oops.


The core of the problem is that a pool is willing to acknowledge a
PoolMigrationUpdate message even though it has an active task on the
file. This patch adds an explicit check for this case and thus prevents
these cyclic tasks.

Notice that although I add a new subclass of CacheException, I do not
actually instantiate it. That is to prevent compatibility issues with
existing pools.


Target: trunk
Target: 1.9.4
Target: 1.9.3
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/399/