build: update dcache.jar when configuration files change
Currently, dcache.jar is only rebuilt when one of the java source files changes. But when using spring cells, a change in the cell configuration files should be reflected in the contents of dcache.jar since that is where those files will be stored. This patch corrects the build file's behaviour by separating the logic that verifies if dcache.jar is up-to-date between the source (java) and configuration (xml and jdo) files present in the source code and rebuilding the jar whenever one of the tests fails, but only starting a re-compilation of sources when former are outdated.

Target: trunk
Acked-by: Tigran
Patch: http://rb.dcache.org/r/497/