core: Include DN and FQAN with subject
This patch extends Subjects.getSubject to include the DN and FQAN if
available. We use GlobusPrincipal to represent the DN and add a custom
FQANPrincipal for the FQAN.

We don't currently use this information, but this patch could be used to
move the tape protection PEP from the doors to the PoolManager.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/502/