Add information about how to decommission the old info provider
The currently available instructions on how to set up the info-cell based info provider fails to mention how to
decommission the old info provider (which was based on the infoProvider cell).  Failure to properly decommission
this old info provider can result in GIP attempting to merge the output from both, the results of which may be
difficult to predict.

This patch adds a short section describing how to decommission the old info provider.  It also adds a short sentence
mentioning that reconfiguring the info provider may result in a short interruption to publishing information (which
came as a surprise to some).

Acked-by: Owen
http://rb.dcache.org/r/404/