Fix bug in info service where secondary information isn't always updated
The information in the info service can be grouped into to classes: primary and secondary.  The primary data is that 
information coming from an authoritative source within dCache; for example, the SrmSpaceManager component provides a 
list of SRM reservations.

Other data can be derived from this primary information; for example, the space information for a poolgroup can be 
derived by summing the same quantities of member pools (the poolgroup total size is the sum of each member pool's 
total size).  This derived information is provided by Secondary Information Providers (SIPs).  The SIPs update 
derived data only when the primary data they use is updated: if an update does not affect the primary data used by a 
SIP then no new derived data is (re-)calculated.

This patch fixes a bug where a SIP was not triggered even when the corresponding primary data was updated.  The 
patch is for all supported branches (1.9.1, 1.9.2, 1.9.3, 1.9.4) and Trunk.

Acked-by: Gerd
http://rb.dcache.org/r/485/