core: Turn Origin into a Principal
The ACL module can authorize based on the origin of a request. That
means the origin is part of the users identity. This patch makes Origin
implement the Principal and Serializable interfaces. Support is added to
the Subjects class to extract the Origin from the subject.

Target: trunk
Acked-by: Irina Kozlova <irina.kozlova@desy.de>
Patch: http://rb.dcache.org/r/509/