pnfsmanager: Add ACL and MODE to FileAttributes
FileAttribute already defined the attribute ACL, however FileAttributes
lacked a field for storing the value. This patch adds the missing field,
and adds support to PnfsManager to query this field through
PnfsGetFileAttributes. chimera.batch and pnfs.batch are updated with the
necessary connection parameters for talking to the ACL database. The
property aclEnabled is used to enable/disable ACL support. When
disabled, querying the ACL property results in a null ACL (equivalent to
an empty ACL).

Several classes in the ACL module have been made Serializable.

The patch also adds support for the MODE FileAttribute.

Target: trunk
Acked-by: Irina Kozlova <irina.kozlova@desy.de>
Patch: http://rb.dcache.org/r/508/