ftp: Use FileAttributes to gather file attributes for listing
Rather than using the mounted file system, or querying for FileMetaData,
this patch updates the FTP list code to use the new FileAttributes
class. The actual listing is still done on the mounted file system.

The current structure is temporary; it is slower than the previous code
as more queries are send to PnfsManager. Once the list support is added
to PnfsManager, this problem will go away.

Target: trunk
Patch: http://rb.dcache.org/r/512/
Acked-by: Marco Devesas Campos <devesas.campos@desy.de>