Add cells jar to satisfy log4j's default usage of PinboardAppender
When running the Chimera CLI, an exception is thrown resulting in a
stacktrace being printed to the console.  This is from the default
log4j file's usage of PinboardAppender.  A fix for this has gone into
Trunk, but not 1.9.4, 1.9.3, 1.9.2 and 1.9.1.

This patch copies this fix to these branches so future release can use
the Chimera CLI without seeing a stacktrace.

http://rb.dcache.org/r/511/
Acked-by: Tigran