core: Add support for running dCache as non-root
This patch adds support for configuring dCache to run as a non-root
user. A new parameter, user, is added to dCacheSetup for specifying the
user name under which to run dCache. When specified, the java process is
executed as that user. The log files are still generated by the user
running the init script (typically root). Thus we can still write log
files to /var/log/ even when running as non-root. PID files are
generated by the user running the JVM, however the scripts take care to
change the owner of the PID file to that user before executing the JVM.
Thus the default location of /var/run/ for PID files still works.

The patch introduces the script share/lib/daemon. It serves a role 
similar to start-stop-daemon on Debian, and the daemon() shell function 
on RedHat: It daemonizes processes that do not have the ability to do 
that (such as the JVM).

A few related changes:

- For logMode being "new", we no longer rename the log file when
  crashing. The log file is only moved when we start dCache, not during
  automatic restart. This avoids that crashes are "lost" due to the log
  file being automatically overwritten. It is also a necessity, as the
  restart is controlled by the unpriviledged user, and that user does
  not have permission to move the log file.

- If config/delay is missing, we now default to a restart delay of 10s,
  thus preventing (or at least throttling) that the disk fills up with
  restart messages in case of a restart loop.

- The 'stop file' has been moved from jobs/ to /tmp. That is because
  writing temporary files to /opt/d-cache/ is considered bad by many
  admins. In particular now that we have the ability to run as a
  non-priviledged user, it makes sense to protect /opt/d-cache/


Just because we now have the ability to start dCache as a non-root user
doesn't mean that dCache will be able to operate as that user. Some
possible sources of failures:

- httpdDomain needs write permission to /opt/d-cache/billing
- all domains talking to a DB need permission to do so
- pool domains need permission to write to the pool directory
- most doors need access to the host certificate and the host key
- gPlazma needs access to /etc/grid-security/
- domains that use toplink attempt to write files like createDDL.jdbc
- currently doors still require read permission to the mounted namespace

Target: trunk
Patch: http://rb.dcache.org/r/529/
Acked-by: Paul Millar <paul.millar@desy.de>