core: Add directory listing to PnfsManager
With this patch the PnfsManager can produce directory listings. The
result is returned in one or more messages depending on the length of
the directory. Breaking up the result avoids excessive memory
consumption in the PnfsManager.

The PnfsManager can optionally apply a filter to the directory (pattern
on file name and a range for bracketing the result). The PnfsManager can
optionally return the FileAttributes of entries in addition to name and
PnfsId.

PnfsManager uses dedicated threads for listing. This is to avoid that
directory listing (which may be expensive) starves other requests.

A client stub, ListDirectoryHandler, is provided to easily request
directory listings. The client stub returns a DirectoryStream object to
avoid building the complete list of directories in memory. Ideally one
would use the same stream interface for the NameSpaceProvider in
PnfsManager, however with the limitted File.listFiles method in current
versions of Java, this would require an additional thread. We should
revisit this once Java 7 is released. Two utility methods for writing to
a DirectoryListPrinter are provided.

The list commands of the FTP door have been updated to use the new
client stub.

The PnfsManager does not throttle how fast it sends reply objects back
to the door. This could potentially cause an OOM condition in the door.
I am not sure how to address this and whether we should address this.

ToDo:

- The Chimera name space provider builds the complete diretory
  representation in memory. This is forced by the interface to Chimera
  and the Chimera lib needs to be updated to fix this. Tigran will do
  this in a separate patch.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/472/