Fix broken example
The pin() call in say() method cannot work as CellAdapter doesn't have this method any more.  Eclipse was picking this up and it was 
bugging me.  The fix also removes some class private fields that are never read.

The change is slightly different from that in the RB since Gerd marked it ship-it but included some comments, which I've tried to 
address.

http://rb.dcache.org/r/577

Acked-by: Gerd