Add support for Jon's p2p dynamic threshold
Jon has a script that monitors the CPU cost for all the pools.  It then ranks these pools based on their CPU cost.  The n-th percentile 
pool cost is chosen, where the n-th percentile is the cost of the pool within that ranking: 0% selects the lowest pool cost, 50% selects 
the median cost and 100% selects the highest pool cost.  This cost is used as the threshold for establishing pool-to-pool "on cost" 
transfers.

The following patch implements this behaviour.  Specifying a on-cost value as a number not ending with "%" will result in the current 
behaviour replicated; all current dCache deployments will have such a value.  Specifying a value ending with "%" will result in the 
percentile cost being calculated dynamically and the resulting value used as the threshold for on-cost pool-to-pool transfers.

The n-th percentile cost is cached.  The cached value is recalculated if a different percentile cost is requested or if a PoolUpMessage 
results in the cache being invalidated.  To prevent all pool-up messages from invalidating the cache, the old- and new- costs are 
recalculated. This check may incur a slight CPU overhead, but this should be outweighed by the saving by avoiding needless recalculation 
of the percentile cost.