It is possible that a StorageInfo object has multiple location URIs that are identical.
Previously, two GenericStorageInfo objects are required to have the same number of URIs.  If the same location URIs are present but in 
differing quantities then the two objects are considered unequal.

This patch allows uneven number of URIs.  If a location URI is present in one object then it must also be present in the other object for 
the two objects to be considered equal, but multiple copies of equal URIs are considered equivalent to having just the one location URI.