PnfsManager: Add timeout aware list partitioning
List partition only took message size into account, not the timeout of
the request. The ListDirectoryHandler employs a timeout between two
replies, not for the complete request. Hence it makes sense for the
PnfsManager to send a reply smaller than the configured size if it knows
that the timeout approaches.

To set the TTL field of the request, ListDirectoryHandler was updated to
use PnfsHandler.pnfsRequest rather than send. The call was already
blocking before, so there is no visible change in behaviour. With this
change we can also provide better error reporting, such as
NoRouteToCellException.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/563/